+++ src/Asv.Mavlink/Protocol/Messages/asv_audio.cs
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-// This code was generate by tool Asv.Mavlink.Shell version 4.0.0-dev.15+a2f1de3777820636a46d83925144e965a9eb2291 25-05-11.
+// This code was generate by tool Asv.Mavlink.Shell version 4.0.0-dev.16+8bb2f8865168bf54d58a112cb63c6bf098479247 25-05-12.
 
 using System;
 using System.Text;
@@ -207,8 +207,8 @@ namespace Asv.Mavlink.AsvAudio
         
         public override void ReadFields(IMavlinkFieldWriter writer)
         {
-            writer.Write(StaticFields[0], Payload.Codec);
-            writer.Write(StaticFields[1], Payload.Mode);
+            writer.Write(StaticFields[0], (ulong)Payload.Codec);
+            writer.Write(StaticFields[1], (ulong)Payload.Mode);
             writer.Write(StaticFields[2], Payload.Name);
         }
